requirements-updated.txt
# Core Flask Dependencies
flask>=3.1.1
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-socketio>=5.5.1
flask-cors>=6.0.1

# Database
psycopg2-binary>=2.9.10

# Environment and Configuration
python-dotenv>=1.1.1

# HTTP and Authentication
requests>=2.32.4
oauthlib>=3.3.1

# Server and Networking
eventlet>=0.40.2
werkzeug>=3.1.3

# Scientific Computing (for AI features)
numpy>=2.2.6
opencv-python>=4.12.0.88
scikit-learn>=1.7.1
pillow>=11.3.0

# Additional Dependencies for Full Functionality
selenium>=4.25.0
email-validator>=2.2.0
python-socketio>=5.13.0
pyopenssl

# Optional but recommended
gunicorn  # for production deployment